+++ spring-pulsar-cache-provider-caffeine/src/main/resources/META-INF/services/org.springframework.pulsar.cache.provider.CacheProviderFactory
@@ -1,0 +1,1 @@
+org.springframework.pulsar.cache.provider.caffeine.CaffeineCacheProviderFactory
