--- spring-pulsar-cache-provider-caffeine/src/main/resources/META-INF/services/org.springframework.pulsar.core.CacheProviderFactory
@@ -1,1 +1,0 @@
-org.springframework.pulsar.core.CaffeineCacheProviderFactory
