+++ spring-pulsar-cache-provider/src/main/java/org/springframework/pulsar/cache/provider/CacheProvider.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.pulsar.core;
+package org.springframework.pulsar.cache.provider;
 
 import java.util.Map;
 import java.util.function.BiConsumer;
