+++ spring-pulsar-cache-provider/src/test/java/org/springframework/pulsar/cache/provider/CacheProviderFactoryTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.pulsar.core;
+package org.springframework.pulsar.cache.provider;
 
 import static org.assertj.core.api.Assertions.assertThat;
