--- spring-pulsar-cache-provider/src/test/resources/META-INF/services/org.springframework.pulsar.core.CacheProviderFactory
@@ -1,1 +1,0 @@
-org.springframework.pulsar.core.CacheProviderFactoryTests$TestCacheProviderFactory
